Moved TableLogger to mttl.logging
