+++ mttl/logging.py
@@ -2,6 +2,11 @@ import logging
 import os
 from functools import lru_cache
 
+import numpy as np
+import pandas as pd
+import prettytable
+import wandb
+
 logger = logging.getLogger("mttl")
 
 
@@ -36,3 +41,42 @@ def setup_logging(log_dir: str = None):
                 "New experiment, log will be at %s",
                 log_file_path,
             )
+
+
+class TableLogger:
+    def __init__(self):
+        self.df = pd.DataFrame()
+
+    def from_df(self, df):
+        self.df = df
+        self.columns = df.columns
+
+    def log(self, row: dict):
+        if self.df is None or len(self.df) == 0:
+            self.df = pd.DataFrame(columns=row.keys())
+        else:
+            # Add new columns to the DataFrame if they don't exist
+            new_columns = set(row.keys()) - set(self.df.columns)
+            for column in new_columns:
+                self.df[column] = np.nan
+        self.df.loc[len(self.df.index)] = row
+
+    def get_table(self):
+        return self.df
+
+    def means(self):
+        # calculate mean for each row, column and diagonal of self.df
+        # filter numeric columns
+        df_numeric = self.df.select_dtypes(include=[np.number])
+        self.df["mean"] = df_numeric.mean(axis=1)
+        self.df.loc["mean"] = df_numeric.mean(axis=0)
+        self.df.loc["mean", "mean"] = np.diag(df_numeric).mean()
+
+    def log_final_table(self):
+        if wandb.run is not None:
+            wandb.log({"table": wandb.Table(data=self.get_table())})
+        table = prettytable.PrettyTable()
+        table.field_names = list(self.df.columns)
+        for i, row in self.df.iterrows():
+            table.add_row(list(row))
+        logger.info("Results:\n" + str(table))
