+++ mttl/utils.py
@@ -5,13 +5,9 @@ import random
 import string
 from typing import Optional
 
-import numpy as np
-import pandas as pd
-import prettytable
 import pytorch_lightning as pl
 import torch
 import torch.nn as nn
-import wandb
 from pytorch_lightning.utilities.rank_zero import rank_zero_only
 
 from mttl.logging import logger
@@ -284,45 +280,6 @@ def rank_zero_only_and_wait(before=True, after=True):
 
 def generate_random_string(str_len=10):
     return "".join(random.choices(string.ascii_uppercase, k=str_len))
-
-
-class TableLogger:
-    def __init__(self):
-        self.df = pd.DataFrame()
-
-    def from_df(self, df):
-        self.df = df
-        self.columns = df.columns
-
-    def log(self, row: dict):
-        if self.df is None or len(self.df) == 0:
-            self.df = pd.DataFrame(columns=row.keys())
-        else:
-            # Add new columns to the DataFrame if they don't exist
-            new_columns = set(row.keys()) - set(self.df.columns)
-            for column in new_columns:
-                self.df[column] = np.nan
-        self.df.loc[len(self.df.index)] = row
-
-    def get_table(self):
-        return self.df
-
-    def means(self):
-        # calculate mean for each row, column and diagonal of self.df
-        # filter numeric columns
-        df_numeric = self.df.select_dtypes(include=[np.number])
-        self.df["mean"] = df_numeric.mean(axis=1)
-        self.df.loc["mean"] = df_numeric.mean(axis=0)
-        self.df.loc["mean", "mean"] = np.diag(df_numeric).mean()
-
-    def log_final_table(self):
-        if wandb.run is not None:
-            wandb.log({"table": wandb.Table(data=self.get_table())})
-        table = prettytable.PrettyTable()
-        table.field_names = list(self.df.columns)
-        for i, row in self.df.iterrows():
-            table.add_row(list(row))
-        logger.info("Results:\n" + str(table))
 
 
 def get_loss(model, evaluator: Evaluator, **kwargs):
